// src/app/models/AbhaData.ts
export interface AbhaData {
    gender: string;
    kycPhoto: string;
    monthOfBirth: string;
    profilePhoto: string;
    stateName: string;    
    city: string;
    dayOfBirth: string;
    yearOfBirth: string;
    pincode: string;
    address: string;
    districtName: string;
    mobile: string;
    firstName: string;
    name: string;
  }
  